+++ common/src/main/java/net/mca/Config.java
@@ -86,6 +86,7 @@ public final class Config implements Serializable {
     public int burnedClothingTickLength = 3600;
     public float coloredHairChance = 0.02f;
     public int heartsRequiredToAutoSpawnGravestone = 10;
+    public boolean enableVillagerChatAI = false;
 
     //village behavior
     public int guardSpawnRate = 6;