+++ common/src/main/java/net/mca/entity/ai/ptg3Modules/PersonalityModule.java
@@ -9,8 +9,8 @@ import static net.mca.entity.ai.PTG3.translate;
 
 public class PersonalityModule {
     public static void apply(List<String> input, VillagerEntityMCA villager, PlayerEntity player) {
-        input.add("$villager is a " + translate(villager.getGenetics().getGender().name()) + ". ");
-        input.add("$villager is " + translate(villager.getVillagerBrain().getMood().getName()) + ". ");
-        input.add("$villager is " + translate(villager.getVillagerBrain().getPersonality().name()) + ". ");
+        input.add("This is a conversation with a " + translate(villager.getGenetics().getGender().name()) + " Minecraft villager named $villager and $player." + " ");
+        input.add("$villager is " + translate(villager.getVillagerBrain().getPersonality().name()) + " and " + translate(villager.getVillagerBrain().getMood().getName()) + ". ");
+        input.add("$villager is a " + translate(villager.getProfession().toString()) + ". ");
     }
 }