+++ changelog.md
@@ -1,6 +1,10 @@
 # 7.4.0
 
 * Added `percentageOfVanillaVillages` as a config value to randomly keep vanilla villages
+* Added special vision for Sirbens
+* Added PTG-3 based villager chat AI
+    * Enable in Config
+    * Speak their name once to start a conversation
 
 # 7.3.21